ZIN-104 Fix flipbook test and fix not working test fail when logger log error etc.

+++ Libs/CoreLib/source/ZtFile.cpp
@@ -56,7 +56,7 @@ namespace zt::core
 				return std::ios_base::in | std::ios_base::out;
 
 			default:
-				Logger->error("Invalid openMode: {}", openMode);
+				Logger->error("Invalid openMode: {}", static_cast<int>(openMode));
 				return std::ios_base::out;
 		}
 	}
